refactor(ai-transport): clarify execution overlay naming
